Streams()

+++ com/eurekaAccounts/stocks/dao/LookUpDAO.java
@@ -61,9 +61,9 @@ public class LookUpDAO extends BaseDAO {
         while (resultSet.next()) {
             StockFundamentalsVO stockFundamentalsVO = new StockFundamentalsVO();
             stockFundamentalsVO.setTickerSymbol(resultSet.getString("ticker_symbol"));
-            stockFundamentalsVO.setSectorId(resultSet.getInt("sector_id"));
+            stockFundamentalsVO.setSectorId(resultSet.getBigDecimal("sector_id"));
             stockFundamentalsVO.setSubSectorId(resultSet.getInt("subsector_id"));
-            stockFundamentalsVO.setMarketCap(resultSet.getLong("market_cap"));
+            stockFundamentalsVO.setMarketCap(resultSet.getBigDecimal("market_cap"));
             stockFundamentalsVO.setCurrentRation(resultSet.getFloat("current_ratio"));
             specificStockFundamentals.add(stockFundamentalsVO);
         }
