+++ com/eurekaAccounts/stocks/dao/LookUpStockFundamentals.java
@@ -34,13 +34,14 @@ public class LookUpStockFundamentals extends BaseDAO {
         while(resultSet.next()){
             StockFundamentalsVO stockFundamentalsVO = new StockFundamentalsVO();
             stockFundamentalsVO.setTickerSymbol(resultSet.getString("ticker_symbol"));
-            stockFundamentalsVO.setSectorId(resultSet.getInt("sector_id"));
+            stockFundamentalsVO.setSectorId(resultSet.getBigDecimal("sector_id"));
             stockFundamentalsVO.setSubSectorId(resultSet.getInt("subsector_id"));
-            stockFundamentalsVO.setMarketCap(resultSet.getLong("market_cap"));
+            stockFundamentalsVO.setMarketCap(resultSet.getBigDecimal("market_cap"));
             stockFundamentalsVO.setCurrentRation(resultSet.getFloat("current_ratio"));
             stockFundamentalsVOS.add(stockFundamentalsVO);
         }
          return stockFundamentalsVOS;
 
     }
+
 }
